Remove PiContext,  now included in PiEditor.
