+++ packages/core/src/editor/PiContext.ts
@@ -1,7 +1,7 @@
 import { PiElement, PiExpression } from "../language";
 
 export interface PiContext {
-    rootElement: PiElement;
+    // rootElement: PiElement;
 
     getPlaceHolderExpression: () => PiExpression;
 }
