--- packages/meta/src/editordef/generator/templates/ContextTemplate.ts
@@ -1,63 +1,0 @@
-import { Names, PathProvider, PROJECTITCORE, LANGUAGE_GEN_FOLDER } from "../../../utils";
-import { PiLanguageUnit } from "../../../languagedef/metalanguage/PiLanguage";
-import { DefEditorLanguage } from "../../metalanguage";
-
-
-export class ContextTemplate {
-    constructor() {
-    }
-
-    generateContext(language: PiLanguageUnit, editorDef: DefEditorLanguage, relativePath: string): string {
-        const rootConceptName = Names.concept(language.rootConcept);
-        const placeHolderConceptName = Names.concept(language.expressionPlaceHolder);
-        const initializationName = Names.initialization(language);
-
-        return `
-            import { action, observable } from "mobx";
-            import { ${Names.PiContext}, ${Names.PiExpression}, ${Names.PiElement}, MobxModelElementImpl } from "${PROJECTITCORE}";
-            import { ${initializationName} } from "../${initializationName}";
-            import { ${rootConceptName} ${(placeHolderConceptName === "" ? "" : ", " + placeHolderConceptName)} } from "${relativePath}${LANGUAGE_GEN_FOLDER }";
-
-            export class ${Names.context(language)} implements PiContext {
-                @observable private _rootElement: PiElement;
-            
-                constructor() {
-                    this.initialize();
-                }
-            
-                set rootElement(exp: PiElement) {
-                    this._rootElement = exp;
-                    if( exp instanceof MobxModelElementImpl) {
-                        exp.container = this;
-                        exp.propertyIndex = undefined;
-                        exp.propertyName = "rootElement";
-                        // not a PiElement , therefore no root.
-                        exp.container = null;
-                    }
-                }
-            
-                get rootElement(): PiElement {
-                    return this._rootElement;
-                }
-            
-                toString(): string {
-                    return "${Names.context(language)}";
-                }
-            
-                ${ !!language.expressionPlaceHolder ? `
-                getPlaceHolderExpression(): PiExpression {
-                    return new ${placeHolderConceptName}; 
-                }`
-            :`  getPlaceHolderExpression(): PiExpression {
-                    return null; 
-                }` }
-            
-                @action
-                private initialize() {
-                    this.rootElement = new ${initializationName}().initialize();
-                }
-            }
-        
-        `;
-    }
-}
