+++ packages/meta/src/editordef/generator/templates/EditorIndexTemplate.ts
@@ -10,7 +10,6 @@ export class EditorIndexTemplate {
         return `
         export * from "./${Names.actions(language)}";
         export * from "./${Names.defaultActions(language)}";
-        export * from "./${Names.context(language)}";
         export * from "./${Names.projectionDefault(language)}";
         `;
     }
