+++ packages/meta/src/editordef/generator/templates/index.ts
@@ -1,5 +1,4 @@
 export * from "./ActionsTemplate";
-export * from "./ContextTemplate";
 export * from "./EditorIndexTemplate";
 export * from "./ProjectionTemplate";
 export * from "./UnparserTemplate";
